+++ src/main/kotlin/lotto/domain/WinningResult.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class WinningResult(lottoList: List<Lotto>, winningLotto: WinningLotto) {
 
@@ -7,7 +7,8 @@ class WinningResult(lottoList: List<Lotto>, winningLotto: WinningLotto) {
     init {
         lottoList.forEach { lotto ->
             val winningNumbers = winningLotto.matchNumbers(lotto)
-            val ranking = RANKING.countOf(winningNumbers.size, lotto.contains(winningLotto.bonusNumber))
+            val ranking =
+                RANKING.countOf(winningNumbers.size, lotto.contains(winningLotto.bonusNumber))
             winningResult[ranking.ordinal] ++
         }
     }
