+++ src/main/kotlin/lotto/domain/WinningStatistics.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import kotlin.math.floor
