+++ src/test/kotlin/lotto/LottoTest.kt
@@ -2,6 +2,14 @@ package lotto
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import lotto.domain.Lotto
+import lotto.domain.LottoMachine
+import lotto.domain.LottoNumber
+import lotto.domain.RANKING
+import lotto.domain.StringNumbers
+import lotto.domain.WinningLotto
+import lotto.domain.WinningResult
+import lotto.domain.WinningStatistics
 import org.assertj.core.api.Assertions
 
 class LottoTest : StringSpec({
@@ -75,16 +83,16 @@ class LottoTest : StringSpec({
         val winningResult = WinningResult(lottoList, winningLotto)
         val winningStatistics = WinningStatistics(lottoMachine.price)
 
-        if (winningStatistics.rateOfReturn(winningResult.getWinningPrice()) < 1) "손해" else "이익" shouldBe "이익"
+        if (winningStatistics.rateOfReturn(winningResult.getWinningPrice()) >= 1) "이익" shouldBe "이익"
     }
 
     "5등." {
         val lottoString = StringNumbers("1,2,3,4,5,6".split(","))
         val lotto = Lotto(lottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
-        val bonusNumber = getBonusNumber(lotto)
 
         val winningLottoString = StringNumbers("1,2,3,7,8,9".split(","))
         val winningLotto = Lotto(winningLottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
+        val bonusNumber = getBonusNumber(winningLotto)
 
         val winningResult = WinningResult(listOf(lotto), WinningLotto(winningLotto, bonusNumber))
 
@@ -94,10 +102,10 @@ class LottoTest : StringSpec({
     "4등." {
         val lottoString = StringNumbers("1,2,3,4,5,6".split(","))
         val lotto = Lotto(lottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
-        val bonusNumber = getBonusNumber(lotto)
 
         val winningLottoString = StringNumbers("1,2,3,4,8,9".split(","))
         val winningLotto = Lotto(winningLottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
+        val bonusNumber = getBonusNumber(winningLotto)
 
         val winningResult = WinningResult(listOf(lotto), WinningLotto(winningLotto, bonusNumber))
 
@@ -107,10 +115,10 @@ class LottoTest : StringSpec({
     "3등." {
         val lottoString = StringNumbers("1,2,3,4,5,6".split(","))
         val lotto = Lotto(lottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
-        val bonusNumber = getBonusNumber(lotto)
 
         val winningLottoString = StringNumbers("1,2,3,4,5,9".split(","))
         val winningLotto = Lotto(winningLottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
+        val bonusNumber = getBonusNumber(winningLotto)
 
         val winningResult = WinningResult(listOf(lotto), WinningLotto(winningLotto, bonusNumber))
 
@@ -120,10 +128,10 @@ class LottoTest : StringSpec({
     "1등." {
         val lottoString = StringNumbers("1,2,3,4,5,6".split(","))
         val lotto = Lotto(lottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
-        val bonusNumber = getBonusNumber(lotto)
 
         val winningLottoString = StringNumbers("1,2,3,4,5,6".split(","))
         val winningLotto = Lotto(winningLottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
+        val bonusNumber = getBonusNumber(winningLotto)
 
         val winningResult = WinningResult(listOf(lotto), WinningLotto(winningLotto, bonusNumber))
 
@@ -133,10 +141,10 @@ class LottoTest : StringSpec({
     "2등." {
         val lottoString = StringNumbers("1,2,3,4,5,6".split(","))
         val lotto = Lotto(lottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
-        val bonusNumber = LottoNumber(6)
 
         val winningLottoString = StringNumbers("1,2,3,4,5,9".split(","))
         val winningLotto = Lotto(winningLottoString.numbers.map { LottoNumber(it.trim()) }.toSet())
+        val bonusNumber = LottoNumber(6)
 
         val winningResult = WinningResult(listOf(lotto), WinningLotto(winningLotto, bonusNumber))
